Create  read_AM2320.py
This is just a prove of concept, so the code is not fault tolerant and the CRC check is missing. You can find CRC verification algorithm  in AM2320 documentation. 
The procedure utilise {smbus} module, can be installed on Rpi by <pip3 install smbus>.
The SM2320 return list of bytes. You can take r[:-2] bytes for CRC verification.